Define base multi-parser support, works as before with Pdf only.
