+++ app/services/parsers.rb
@@ -1,0 +1,5 @@
+module Parsers
+  def self.parser_for(filename)
+    return Parsers::Pdf if filename.end_with?(".pdf")
+  end
+end
